# src/openapi.yaml
openapi: '3.0.3'
info:
  title: FUME Community API
  version: 0.0.0
  description: |
    FHIR-Utilized Mapping Engine — Community Edition.  
    Evaluate ad-hoc FUME expressions or execute saved StructureMap mappings.
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
...
servers:
  - url: /
    description: Current server
components:
  schemas:
...
    RecacheResponse:
      type: object
      description: Recache result describing what was loaded into cache.
      properties:
        message:
          type: string
          example: The following Mappings were loaded to cache
        mappings:
          type: array
          items:
            type: string
          description: List of mapping keys loaded into cache
      example:
        message: The following Mappings were loaded to cache
        mappings:
          - Patient12
          - exampleForRepo
...
paths:
...
  /$recache:
    post:
      tags:
        - Cache
      summary: Reload mappings
      description: >-
        Reloads mappings and aliases from the configured sources into cache.
        Flushes the ConceptMap cache. Available only when at least one mapping
        source is configured.
      operationId: recache
      responses:
        '200':
          description: Mappings reloaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecacheResponse'
        '405':
          description: Method not allowed. No mapping sources configured.
        '500':
          description: Cache could not be refreshed
  /recache:
    get:
      tags:
        - Cache
      summary: Recache (method not allowed)
      description: >-
        Recache has side-effects, so GET is no longer supported. Use `POST
        /$recache` instead.
      operationId: recacheGetDisabled
      deprecated: true
      responses:
        '405':
          description: Method not allowed. Use POST /$recache.
    post:
      tags:
        - Cache
      summary: Reload mappings (deprecated)
      description: >-
        Same as `POST /$recache`, but deprecated. Emits a warning in logs and a
        Warning response header. Will be removed in future releases.
      operationId: recacheDeprecated
      deprecated: true
      responses:
        '200':
          description: Mappings reloaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecacheResponse'
        '405':
          description: Method not allowed. No mapping sources configured.
        '500':
          description: Cache could not be refreshed
...